[UPD] #158 combine slur and slur_commands cogs, update slur detection related commands
