+++ utils/detection.py
@@ -32,6 +32,10 @@ def _leet(word: str) -> list[str]:
     return ["".join(permutations) for permutations in product(*possibles)]
 
 
+def clear_string(text: str) -> str:
+    return _special_regex.sub("", text.lower())
+
+
 def highlight_matches(text: str, matches: list[re.Match]) -> str | list[str]:
     highlighted = ""
     pos = 0
@@ -48,7 +52,7 @@ def highlight_matches(text: str, matches: list[re.Match]) -> str | list[str]:
 @dataclass
 class SlurPattern:
     slur: str
-    goodwords: list[str]
+    goodwords: set[str]
 
     def __post_init__(self):
         self._create_detector_regex()
@@ -83,9 +87,10 @@ class SlurPattern:
 
 
 class SlurDetector:
-    slurs: set[SlurPattern]
+    slurs: dict[str, SlurPattern]
     regex: re.Pattern
     unleet: dict[str, str]
+    goodwords: set[str]
 
     def __init__(self):
         self.load()
@@ -95,26 +100,48 @@ class SlurDetector:
             slurs = session.query(Slur).all()
             goodwords = session.query(Goodword).all()
 
-            slur_patterns = [
-                SlurPattern(
+            self.slurs = {
+                slur.slur: SlurPattern(
                     slur.slur,
-                    [
+                    {
                         goodword.goodword
                         for goodword in goodwords
                         if goodword.slur == slur.slur
-                    ],
+                    },
                 )
                 for slur in slurs
-            ]
+            }
 
-            self.slurs = set(slur_patterns)
-            self.regex = re.compile(
-                "|".join([slur.regex.pattern for slur in self.slurs]), re.IGNORECASE
-            )
+            self._create_regex()
 
             self.unleet = {}
-            for slur in self.slurs:
+            for slur in self.slurs.values():
                 self.unleet.update({leet: slur.slur for leet in _leet(slur.slur)})
+
+    def _create_regex(self):
+        self.regex = re.compile(
+            "|".join([slur.regex.pattern for slur in self.slurs.values()]),
+            re.IGNORECASE,
+        )
+
+    def add_slur(self, slur: str):
+        self.slurs[slur] = SlurPattern(slur, [])
+        self._create_regex()
+        self.unleet.update({leet: slur for leet in _leet(slur)})
+
+    def remove_slur(self, slur: str):
+        self.slurs.pop(slur)
+        self._create_regex()
+
+    def add_goodword(self, slur: str, goodword: str):
+        slur: SlurPattern = self.slurs[slur]
+        slur.goodwords.add(goodword)
+        slur._create_goodword_regex()
+
+    def remove_goodword(self, slur: str, goodword: str):
+        slur: SlurPattern = self.slurs[slur]
+        slur.goodwords.remove(goodword)
+        slur._create_goodword_regex()
 
     def find_slurs(self, __value: str) -> dict[str, list[re.Match]]:
         slurs: dict[str, list] = {}
@@ -128,7 +155,7 @@ class SlurDetector:
             pos = match.end()
 
         # Remove matches that are part of goodword
-        for slur in self.slurs:
+        for slur in self.slurs.values():
             if slur.slur not in slurs or not slur.goodwords:
                 continue
             pos = 0
@@ -144,6 +171,15 @@ class SlurDetector:
                 pos = match.end()
 
         return {slur: matches for slur, matches in slurs.items() if matches}
+
+    def __getattr__(self, __name: str) -> set[str]:
+        if __name == "goodwords":
+            return {
+                goodword
+                for slur in self.slurs.values()
+                for goodword in slur.goodwords
+            }
+        raise AttributeError(f"'{self.__class__.__name__}' object has no attribute '{__name}'")
 
     def __contains__(self, __value: str) -> bool:
         return self.regex.search(__value) is not None
